+++ ricaun.Revit.UI.Example/Proprieties/Resource.Designer.cs
@@ -63,9 +63,9 @@ namespace ricaun.Revit.UI.Example.Proprieties {
         /// <summary>
         ///   Looks up a localized resource of type System.Drawing.Bitmap.
         /// </summary>
-        internal static System.Drawing.Bitmap icon {
+        internal static System.Drawing.Bitmap LargeImage {
             get {
-                object obj = ResourceManager.GetObject("icon", resourceCulture);
+                object obj = ResourceManager.GetObject("LargeImage", resourceCulture);
                 return ((System.Drawing.Bitmap)(obj));
             }
         }