+++ ricaun.Revit.UI.Example/Revit/App.cs
@@ -1,20 +1,27 @@
 using Autodesk.Revit.DB;
 using Autodesk.Revit.UI;
-using ricaun.Revit.UI;
+using Autodesk.Revit.ApplicationServices;
 using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using System.Windows.Media;
 
 namespace ricaun.Revit.UI.Example.Revit
 {
     [Console]
     public class App : IExternalApplication
     {
+        private const string TabName = "ricaun";
+        private const string PanelName = "Example";
         private static RibbonPanel ribbonPanel;
+
         public Result OnStartup(UIControlledApplication application)
         {
-            ribbonPanel = application.CreatePanel("Example");
-            ribbonPanel.AddPushButton<Commands.Command>();
+            ribbonPanel = application.CreatePanel(TabName, PanelName);
+            var button = ribbonPanel.AddPushButton<Commands.Command>();
 
-            ribbonPanel.CreatePulldownButton(new[] {
+            ribbonPanel.CreatePulldownButton("PulldownButton", new[] {
                 ribbonPanel.NewPushButtonData<Commands.Command>(),
                 ribbonPanel.NewPushButtonData<Commands.Command>(),
                 ribbonPanel.NewPushButtonData<Commands.Command>(),
@@ -23,7 +30,7 @@ namespace ricaun.Revit.UI.Example.Revit
                 ribbonPanel.NewPushButtonData<Commands.Command>(),
                 ribbonPanel.NewPushButtonData<Commands.Command>() });
 
-            ribbonPanel.CreateSplitButton(new[] {
+            ribbonPanel.CreateSplitButton("SplitButton", new[] {
                 ribbonPanel.NewPushButtonData<Commands.Command>(),
                 ribbonPanel.NewPushButtonData<Commands.Command>(),
                 ribbonPanel.NewPushButtonData<Commands.Command>(),
@@ -33,7 +40,7 @@ namespace ricaun.Revit.UI.Example.Revit
                 ribbonPanel.NewPushButtonData<Commands.Command>(),
                 ribbonPanel.NewPushButtonData<Commands.Command>() });
 
-            ribbonPanel.AddPushButton<Commands.Command>("");
+            var button2 = ribbonPanel.AddPushButton<Commands.Command>("");
 
             var items = ribbonPanel.AddStackedItems(
                 ribbonPanel.NewPushButtonData<Commands.Command>("Item1"),
@@ -41,9 +48,8 @@ namespace ricaun.Revit.UI.Example.Revit
 
             foreach (var item in items)
             {
-                var ri = item.GetRibbonItem();
-                ri.ShowText = false;
-                ri.Size = Autodesk.Windows.RibbonItemSize.Large;
+                item.SetItemSize();
+                item.SetText();
             }
 
             var item3s = ribbonPanel.AddStackedItems(
@@ -53,23 +59,22 @@ namespace ricaun.Revit.UI.Example.Revit
 
             foreach (var item in item3s)
             {
-                var ri = item.GetRibbonItem();
-                ri.ShowText = false;
+                item.SetText();
             }
-
+            /*
             foreach (var item in ribbonPanel.GetItems())
             {
                 if (item is PushButton pushButton)
                 {
-                    pushButton.LargeImage = Proprieties.Resource.icon.GetBitmapSource();
-                    pushButton.Image = Proprieties.Resource.icon.GetBitmapSource().Scale(0.5);
-                    pushButton.ToolTipImage = Proprieties.Resource.icon.GetBitmapSource().Scale(2);
+                    pushButton.LargeImage = Proprieties.Resource.LargeImage.GetBitmapSource();
+                    pushButton.Image = Proprieties.Resource.LargeImage.GetBitmapSource().Scale(0.5);
+                    pushButton.ToolTipImage = Proprieties.Resource.LargeImage.GetBitmapSource().Scale(2);
                 }
                 if (item is PulldownButton pulldownButton)
                 {
-                    pulldownButton.LargeImage = Proprieties.Resource.icon.GetBitmapSource();
-                    pulldownButton.Image = Proprieties.Resource.icon.GetBitmapSource().Scale(0.5);
-                    pulldownButton.ToolTipImage = Proprieties.Resource.icon.GetBitmapSource().Scale(2);
+                    pulldownButton.LargeImage = Proprieties.Resource.LargeImage.GetBitmapSource();
+                    pulldownButton.Image = Proprieties.Resource.LargeImage.GetBitmapSource().Scale(0.5);
+                    pulldownButton.ToolTipImage = Proprieties.Resource.LargeImage.GetBitmapSource().Scale(2);
                 }
                 if (item is SplitButton splitButton)
                 {
@@ -77,13 +82,63 @@ namespace ricaun.Revit.UI.Example.Revit
                     {
                         if (i is PushButton pb)
                         {
-                            pb.LargeImage = Proprieties.Resource.icon.GetBitmapSource();
-                            pb.Image = Proprieties.Resource.icon.GetBitmapSource().Scale(0.5);
-                            pb.ToolTipImage = Proprieties.Resource.icon.GetBitmapSource().Scale(2);
+                            pb.LargeImage = Proprieties.Resource.LargeImage.GetBitmapSource();
+                            pb.Image = Proprieties.Resource.LargeImage.GetBitmapSource().Scale(0.5);
+                            pb.ToolTipImage = Proprieties.Resource.LargeImage.GetBitmapSource().Scale(2);
                         }
                     }
                 }
             }
+            */
+
+            ribbonPanel.GetRibbonPanel().Tab.SetOrderPanels();
+            var ric = ribbonPanel.GetRibbonPanel().Tab.Panels.ToList().FirstOrDefault(e => e.Source.Title == "ricaun");
+            ric?.MoveRibbonPanel();
+
+            foreach (var item in ribbonPanel.GetRibbonItems())
+            {
+                //Console.WriteLine($"{item} {item.Name}");
+            }
+
+            ribbonPanel.UpdateRibbonDescription(setting =>
+            {
+                setting.Add("",
+                    new RibbonDescription()
+                    {
+                        LargeImage = Proprieties.Resource.LargeImage.GetBitmapSource(),
+                        Help = "https://ricaun.com"
+                    }
+                );
+
+                setting.Add<Commands.Command>(
+                    new RibbonDescription()
+                    {
+                        Text = "Hello",
+                        ToolTip = "This is a Tool Tip",
+                        LongDescription = "This is a Long Description",
+                        LargeImage = Proprieties.Resource.LargeImage.GetBitmapSource(),
+                        Help = "https://ricaun.com"
+                    },
+                    new RibbonDescription(LanguageType.Brazilian_Portuguese)
+                    {
+                        Text = "Ola",
+                        ToolTip = "Este é um Tool Tip",
+                        LongDescription = "Este é um Long Description",
+                    }
+                );
+
+                setting.Add("PulldownButton", new RibbonDescription()
+                {
+                    Text = "PulldownButton",
+                    Help = "https://ricaun.com"
+                });
+
+                setting.Add("SplitButton", new RibbonDescription()
+                {
+                    Text = "SplitButton",
+                    Help = "https://ricaun.com"
+                });
+            });
 
             return Result.Succeeded;
         }
@@ -93,9 +148,5 @@ namespace ricaun.Revit.UI.Example.Revit
             ribbonPanel.Close();
             return Result.Succeeded;
         }
-    }
-
-    internal class ConsoleAttribute : Attribute
-    {
     }
 }